Trait emu_core::cache::Cache

pub trait Cache {
    fn contains(key: u64) -> bool;

    fn get(key: u64) -> Arc<DeviceFnMut>;

    fn insert(key: u64, device_fn_mut: Arc<DeviceFnMut>);
}

A trait to implement to create your own cache

The compile function is generic over a Cache implementation. So you could, for example, implement Cache for a disk cache or in-memory cache customized for your needs.

Required methods

fn contains(key: u64) -> bool

fn get(key: u64) -> Arc<DeviceFnMut>

fn insert(key: u64, device_fn_mut: Arc<DeviceFnMut>)

Loading content...