Module emu_core::compile

The whole source-to-DeviceFnMut compilation pipeline

Structs

Spirv
A wrapper for SPIR-V bytecode

SpirvBuilder
A builder for constructing a Spirv

Enums

SpirvOrFinished
Either a finished DeviceFnMut or compiled SPIR-V

Traits

CompileToSpirv
The trait to implement when adding support for a new source language (e.g. - HLSL, XLA, Swift SIL, etc.).

GlslStruct
A trait for structures that can exist in both Rust and GLSL

Functions

compile
Compiles the given source to SpirvOrFinished