Trait emu_core::compile::GlslStruct

pub trait GlslStruct {
    fn as_glsl() -> String;
}

A trait for structures that can exist in both Rust and GLSL

Required methods

fn as_glsl() -> String

Provides the GLSL structure definition code to define this structure in GLSL

Loading content...