Module emu_core::device

The lowest-level, core functionality for controlling a GPU device

You can use this module in isolation from the rest of Emu for the most fine control over the internals of the library. The most important things in this module that are likely of interest to you are Device, DeviceBox<T> , and DeviceFnMut.

Structs

ArgAndParamInfo
Information that is held by both arguments and by parameters

ArgsBuilder
Helps with building a DeviceFnMutArgs

Device
Represents a single device

DeviceBox
A type for boxing stuff stored on a device

DeviceFnMut
Represents a compiled kernel that can then be launched across spawned threads with Device::call or spawn

DeviceFnMutArgs
Holds the actual arguments to be passed into a DeviceFnMut

DeviceFnMutParams
Describes the parameters that can be passed to a DeviceFnMut

DeviceInfo
Contains information about a device

ParamsBuilder
Helps with building a DeviceFnMutParams

Enums

DeviceType
Represents a type of device

Mutability
Says whether or not something is mutable

Functions

convert_to_spirv
Converts a slice of bytes to a slice of 4-byte words