Module emu_core::error

Various error types

Structs

CompileError
An error for compilation failures

CompletionError
An error for failure to complete data movement or computation

NoDeviceError
An error for when there is no device to complete a certain operation

PoolAlreadyInitializedError
An error that occurs when you attempt to initialize an already initialized pool of devices

RuntimeError
A runtime error that occurs on the device

UnavailableDeviceError
An error when a device is there but not available for use

Enums

CompileOrNoDeviceError
An error for capturing compilation fails or no device present

GetError
An error in getting data stored in a DeviceBox

LaunchError
An error in launching kernels