Module emu_core::pool

Tools for managing the device pool either implicitly or explicitly

If you plan on not really delving into the WebGPU internals of Emu and just want to work from the other side of the abstraction, the only important thing here is assert_device_pool_initialized.

Structs

DevicePoolMember
Represents a member of the device pool

DevicePoolMemberInfo
Holds information about a member of the device pool

Functions

assert_device_pool_initialized
Asserts that the device pool has been initialized

info
Returns information about the currently selected device

info_all
Returns information about all devices in the pool

pool
Sets the device pool to the given Vec of devices

select
Selects a device from the pool using the given selector function

take
Takes the device currently selected out of the device pool and hands you a mutex for mutating the device's sate